Split out RequestDefinition struct
This moves the `RequestDefinition` struct into its own file.
